frontend/requirements.txt
streamlit
#==1.11.1
requests
#==2.28.1
Pillow
#==9.2.0
numpy
#==1.22.4
pathlib